Document unified React app
