+++ material-kit-react/.env.sample
@@ -1,0 +1,1 @@
+REACT_APP_MESHY_KEY=your_meshy_api_key
