+++ material-kit-react/src/components/Preview.js
@@ -1,0 +1,17 @@
+import React from "react";
+
+const Preview = ({ thumbnailUrl }) => {
+  return (
+    <div>
+      {thumbnailUrl && (
+        <img
+          src={thumbnailUrl}
+          alt="3D Model Preview"
+          style={{ maxWidth: "300px", marginTop: "10px" }}
+        />
+      )}
+    </div>
+  );
+};
+
+export default Preview;
