+++ material-kit-react/src/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import * as ReactDOMClient from "react-dom/client";
-import { BrowserRouter } from "react-router-dom";
+
 import { CssBaseline, ThemeProvider } from "@mui/material";
 import App from "./App"; // Assuming App.js is in the same folder
 import theme from "./theme"; // Import the theme from theme.js
@@ -13,12 +13,11 @@ if (container) {
   const root = ReactDOMClient.createRoot(container);
 
   root.render(
-    <BrowserRouter>
+      
       <ThemeProvider theme={theme}>
         <CssBaseline />
         <App />
       </ThemeProvider>
-    </BrowserRouter>
   );
 } else {
   console.error("Root container not found in the DOM.");
